\name{BivarP-package}
\alias{BivarP-package}
\docType{package}
\title{
Estimating the Parameters of Some Bivariate Distributions
}
\description{
Parameter estimation of bivariate distribution functions
    modeled as a Archimedean copula function. The input data may contain
    values from right censored. Used marginal distributions are two-parameter.
    Methods for density, distribution, survival, random sample generation.
}
\details{
\tabular{ll}{
Package: \tab BivarP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-17\cr
License: \tab GPL (>= 3)\cr
}
}
\author{
Josef Brejcha

Maintainer: Josef Brejcha <brchjo@gmail.com>
}
\references{
M. Mahfoud, "Bivariate Archimedean copulas: an application to two stock market
indices", \emph{Vrije Universiteit Amsterdam}, \bold{BMI Paper}, Amsterdam-2012,
\url{https://www.few.vu.nl/en/Images/werkstuk-mahfoud_tcm39-277460.pdf}
}
\keyword{ package }

